pyproject.toml
[project]
name = "knn-car-reidentification"
version = "0.1.0"
description = "Project to KNN (FIT BUT)"
authors = [
  { name = "Jiří Stehlík", email = "xstehl23@vutbr.cz" },
  { name = "Filip Černý", email = "xcerny81@vutbr.cz" },
  { name = "Marek Halamka", email = "xhalam16@vutbr.cz" },
]
readme = "README.md"
requires-python = ">=3.13, <3.15"
dependencies = ["Pillow"]

[tool.poetry]
package-mode = false

[tool.poetry.group.system-deps.dependencies]
torch = ">=2.0.0"
torchvision = ">=0.20.0"
timm = ">=1.0.10"
pytorch-metric-learning = ">=2.8.0"
pytorch-lightning = ">=2.5.1"

[dependency-groups]
dev = ["ruff (>=0.15.10,<0.16.0)"]

[tool.ruff]
exclude = [
  ".bzr",
  ".direnv",
  ".eggs",
  ".git",
  ".git-rewrite",
  ".hg",
  ".ipynb_checkpoints",
  ".mypy_cache",
  ".nox",
  ".pants.d",
  ".pyenv",
  ".pytest_cache",
  ".pytype",
  ".ruff_cache",
  ".svn",
  ".tox",
  ".venv",
  ".vscode",
  "__pypackages__",
  "_build",
  "buck-out",
  "build",
  "dist",
  "node_modules",
  "site-packages",
  "venv",
]

line-length = 88
indent-width = 4

target-version = "py313"

[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []

fixable = ["ALL"]
unfixable = []

dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

[tool.ruff.format]
quote-style = "double"

indent-style = "space"

skip-magic-trailing-comma = false

line-ending = "auto"

docstring-code-format = true

docstring-code-line-length = "dynamic"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
